@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <Nullable>disable</Nullable>
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="EPiServer.CMS" Version="12.3.2" />
-    <PackageReference Include="Wangkanai.Detection" Version="2.0.1" />
+    <PackageReference Include="EPiServer.CMS" Version="12.4.0" />
+    <PackageReference Include="Wangkanai.Detection" Version="5.2.0" />
   </ItemGroup>
   
   <ItemGroup>
