@@ -1,11 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <Nullable>enable</Nullable>
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="EPiServer.CMS" Version="12.3.2" />
+    <PackageReference Include="EPiServer.CMS" Version="12.4.0" />
   </ItemGroup>
   
   <ItemGroup>
