@@ -22,14 +22,14 @@ namespace MyAppNamespace
             IsStoppable = true;
         }
 
-         /// <summary>
+        /// <summary>
         /// Called when a scheduled job executes
         /// </summary>
         /// <returns>A status message to be stored in the database log and visible from admin mode</returns>
         public override string Execute()
         {
             //Call OnStatusChanged to periodically notify progress of job for manually started jobs
-            OnStatusChanged(String.Format("Starting execution of {0}", this.GetType()));
+            OnStatusChanged($"Starting execution of {GetType()}");
 
             //Add implementation
 
