@@ -1,7 +1,8 @@
 using EPiServer.Web;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
-using Wangkanai.Detection;
+using Wangkanai.Detection.Models;
+using Wangkanai.Detection.Services;
 
 namespace Alloy._1.Business.Channels
 {
@@ -18,8 +19,8 @@ namespace Alloy._1.Business.Channels
 
         public override bool IsActive(HttpContext context)
         {
-            var detection = context.RequestServices.GetRequiredService<IDetection>();
-            return detection.Device.Type == DeviceType.Mobile;
+            var detection = context.RequestServices.GetRequiredService<IDetectionService>();
+            return detection.Device.Type == Device.Mobile;
         }
     }
 }
